requirements.txt
fastapi==0.109.0
flask==3.0.1
pydantic-core==2.14.6
pydantic==2.5.3
PyJWT==2.8.0
requests==2.31.0
uvicorn==0.27.0
